Move concrete providers to their own directory
